requirements.txt
# 项目依赖
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-text-splitters==0.0.1
langchain-openai==0.0.5
chromadb==0.4.24
python-dotenv==1.0.1
pymupdf==1.23.8
unstructured==0.11.2
python-magic==0.4.27
streamlit==1.31.1
openai==1.12.0
deepseek-ai==0.0.1
numpy==1.26.4
pypdf==4.1.0
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.3
zhipuai==1.0.7
jq==1.6.0
pandas==2.1.4

# 测试依赖
pytest==8.0.0
pytest-cov==4.1.0 